requirements.txt
Django==3.1.7
djangorestframework==3.12.2
Pillow==8.1.2
psycopg2-binary==2.8.3
djangorestframework-simplejwt
PyJWT==v1.7.1
python-decouple
django-cors-headers